plugin/package: move in openapischema, use updated buildinfo, go mod init
