+++ cmd/cli/plugin/package/openapischema/schema_default.go
