+++ cmd/cli/plugin/package/openapischema/schema_default_test.go
