+++ cmd/cli/plugin/package/openapischema/suite_test.go
