+++ cmd/cli/plugin/package/openapischema/testdata/contour-data-values.yaml
