+++ cmd/cli/plugin/package/openapischema/testdata/contourschema.yaml
