+++ cmd/cli/plugin/package/openapischema/testdata/grafana-data-values.yaml
