+++ cmd/cli/plugin/package/openapischema/testdata/grafanaschema.yaml
